// src/app/shared/interface/contact-options.interface.ts
import { OptionInterface } from './option.interface';

export interface ContactOptionsInterface {
  language: OptionInterface[];
  title: OptionInterface[];
  greeting: OptionInterface[];
  nationality: OptionInterface[];
  country: OptionInterface[];
  maritalStatus: OptionInterface[];
  children: OptionInterface[];
  pipelineStage: OptionInterface[];
  socialMedia: OptionInterface[];
}